+++ cdap-master/src/main/java/co/cask/cdap/data/tools/UpgradeTool.java
@@ -57,7 +57,6 @@ import co.cask.cdap.explore.guice.ExploreClientModule;
 import co.cask.cdap.internal.app.runtime.schedule.store.ProgramScheduleStoreDataset;
 import co.cask.cdap.internal.app.store.DefaultStore;
 import co.cask.cdap.logging.guice.KafkaLogAppenderModule;
-import co.cask.cdap.logging.meta.LoggingStoreTableUtil;
 import co.cask.cdap.messaging.guice.MessagingClientModule;
 import co.cask.cdap.messaging.store.hbase.HBaseTableFactory;
 import co.cask.cdap.security.auth.context.AuthenticationContextModules;
@@ -461,9 +460,6 @@ public class UpgradeTool {
     }
 
     DefaultStore.setupDatasets(datasetFramework);
-    // logs metadata
-    LoggingStoreTableUtil.setupDatasets(datasetFramework);
-
     ProgramScheduleStoreDataset.setupDatasets(datasetFramework);
   }
 }