+++ cdap-watchdog/src/main/java/co/cask/cdap/logging/framework/AbstractAppenderContext.java
@@ -16,22 +16,10 @@
 
 package co.cask.cdap.logging.framework;
 
-import co.cask.cdap.api.Transactional;
-import co.cask.cdap.api.TxRunnable;
-import co.cask.cdap.api.dataset.DatasetManager;
 import co.cask.cdap.api.logging.AppenderContext;
 import co.cask.cdap.api.metrics.MetricsCollectionService;
 import co.cask.cdap.api.metrics.MetricsContext;
-import co.cask.cdap.data.dataset.SystemDatasetInstantiator;
-import co.cask.cdap.data2.datafabric.dataset.DefaultDatasetManager;
-import co.cask.cdap.data2.dataset2.DatasetFramework;
-import co.cask.cdap.data2.dataset2.MultiThreadDatasetCache;
-import co.cask.cdap.data2.transaction.Transactions;
-import co.cask.cdap.proto.id.NamespaceId;
-import com.google.common.collect.ImmutableMap;
-import org.apache.tephra.RetryStrategies;
-import org.apache.tephra.TransactionFailureException;
-import org.apache.tephra.TransactionSystemClient;
+import co.cask.cdap.spi.data.transaction.TransactionRunner;
 import org.apache.twill.filesystem.LocationFactory;
 
 import java.util.Collections;
@@ -41,47 +29,26 @@ import java.util.Collections;
  */
 public abstract class AbstractAppenderContext extends AppenderContext {
 
-  private final Transactional transactional;
   private final LocationFactory locationFactory;
-  private final DatasetManager datasetManager;
+  private final TransactionRunner transactionRunner;
   private final MetricsContext metricsContext;
 
-  protected AbstractAppenderContext(DatasetFramework datasetFramework,
-                                    TransactionSystemClient txClient,
+  protected AbstractAppenderContext(TransactionRunner transactionRunner,
                                     LocationFactory locationFactory,
                                     MetricsCollectionService metricsCollectionService) {
     this.locationFactory = locationFactory;
-    // No need to have retry for dataset admin operations
-    // The log framework itself will perform retries and state management
-    this.datasetManager = new DefaultDatasetManager(datasetFramework, NamespaceId.SYSTEM,
-                                                    co.cask.cdap.common.service.RetryStrategies.noRetry(), null);
-    this.transactional = Transactions.createTransactionalWithRetry(
-      Transactions.createTransactional(new MultiThreadDatasetCache(
-        new SystemDatasetInstantiator(datasetFramework), txClient,
-        NamespaceId.SYSTEM, ImmutableMap.<String, String>of(), null, null)),
-      RetryStrategies.retryOnConflict(20, 100)
-    );
-    this.metricsContext = metricsCollectionService.getContext(Collections.<String, String>emptyMap());
+    this.transactionRunner = transactionRunner;
+    this.metricsContext = metricsCollectionService.getContext(Collections.emptyMap());
   }
 
   @Override
-  public final DatasetManager getDatasetManager() {
-    return datasetManager;
+  public final TransactionRunner getTransactionRunner() {
+    return transactionRunner;
   }
 
   @Override
   public final LocationFactory getLocationFactory() {
     return locationFactory;
-  }
-
-  @Override
-  public final void execute(TxRunnable runnable) throws TransactionFailureException {
-    transactional.execute(runnable);
-  }
-
-  @Override
-  public final void execute(int timeoutInSeconds, TxRunnable runnable) throws TransactionFailureException {
-    transactional.execute(timeoutInSeconds, runnable);
   }
 
   @Override